Tuning hyperparameter vocab_size  in model.py
The size of the vocabulary used by the model. This determines the number of unique tokens (words, subwords, characters, or other units) the model can understand. For example, if the tokenizer maps the text into 10,000 unique tokens, vocab_size would be 10,000. For a academic project like use 30k to 50k is the ideal size