Update model.py
The dimensionality of the embedding vector for each token. Each token in the input sequence is mapped to a vector of size n_embd, which represents its numerical feature encoding. Larger values allow for richer representations but increase computational cost. Because of the resource constraints we can have an embedding size up to 512